test notebook
the test notebook is testing all steps implemented to convert rover's view into a world map so as to be able to explore an unknown environment and chart a world view map for it.

test images and csv files are used, and an movie about the rover exploring the envronment is generated.